SEQNG-411 Updated to sbt 1.0.4
