+++ modules/edu.gemini.seqexec.model/src/main/scala/edu/gemini/seqexec/model/ActionType.scala
@@ -13,12 +13,12 @@ import scalaz.Equal
 sealed trait ActionType
 
 object ActionType {
-  
+
   object Observe extends ActionType
   // Used in tests
   object Undefined extends ActionType
   final case class Configure(sys: Resource) extends ActionType
-  
+
   implicit val equal: Equal[ActionType] = Equal.equalA[ActionType]
 
 }
